% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_tdt_block}
\alias{as_tdt_block}
\title{Coerce to a materialized TDT block}
\usage{
as_tdt_block(x, ...)
}
\arguments{
\item{x}{A \code{tdt_block}, \code{tdt_block_py}, or compatible object.}

\item{...}{Passed to \code{collect_block()}.}
}
\value{
A materialized \code{tdt_block}.
}
\description{
Coerce to a materialized TDT block
}
