% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as_tibble_epocs}
\alias{as_tibble_epocs}
\title{Return epocs/events as a tibble}
\usage{
as_tibble_epocs(x, store = NULL)
}
\arguments{
\item{x}{A \code{tdt_block}, \code{tdt_block_py}, or compatible object.}

\item{store}{Optional epoc store name.}
}
\value{
A tibble with \code{store}, \code{onset}, \code{offset}, and \code{value} columns.
}
\description{
Return epocs/events as a tibble
}
