% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{block_info}
\alias{block_info}
\title{Return block metadata}
\usage{
block_info(x)
}
\arguments{
\item{x}{A \code{tdt_block}, \code{tdt_block_py}, or compatible object.}
}
\value{
A named list.
}
\description{
Return block metadata
}
