% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_epocs}
\alias{collect_epocs}
\title{Collect epocs/events into R}
\usage{
collect_epocs(x, store = NULL, as = c("tibble", "list"))
}
\arguments{
\item{x}{A \code{tdt_block}, \code{tdt_block_py}, or compatible object.}

\item{store}{Optional epoc store name. If \code{NULL}, all epocs are collected.}

\item{as}{Return shape: \code{"tibble"} or \code{"list"}.}
}
\value{
A tibble or named list of tibbles.
}
\description{
Collect epocs/events into R
}
