% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{epoc}
\alias{epoc}
\title{Return one epoc/event store}
\usage{
epoc(x, store)
}
\arguments{
\item{x}{A \code{tdt_block}, \code{tdt_block_py}, or compatible object.}

\item{store}{Epoc store name.}
}
\value{
One epoc object or tibble subset.
}
\description{
Return one epoc/event store
}
