% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{epoc_filter_py}
\alias{epoc_filter_py}
\title{Filter TDT epocs through Python \code{tdt}}
\usage{
epoc_filter_py(
  data,
  epoc,
  values = NULL,
  modifiers = NULL,
  t = NULL,
  tref = FALSE,
  keepdata = TRUE,
  ...
)
}
\arguments{
\item{data}{A \code{tdt_block_py} wrapper or raw Python block object.}

\item{epoc}{Epoc store name.}

\item{values, modifiers, t, tref, keepdata}{Arguments passed to Python
\code{tdt.epoc_filter()}.}

\item{...}{Additional keyword arguments passed to Python \code{tdt.epoc_filter()}.}
}
\value{
A \code{tdt_block_py} wrapper.
}
\description{
Filter TDT epocs through Python \code{tdt}
}
