% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{epoc_names}
\alias{epoc_names}
\title{Return epoc/event store names}
\usage{
epoc_names(x)
}
\arguments{
\item{x}{A \code{tdt_block}, \code{tdt_block_py}, epoc container, or compatible object.}
}
\value{
A character vector of epoc store names.
}
\description{
Return epoc/event store names
}
