% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{is_tdt_block}
\alias{is_tdt_block}
\title{Test whether an object is a materialized TDT block}
\usage{
is_tdt_block(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{tdt_block}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a materialized TDT block
}
