% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{new_tdt_block}
\alias{new_tdt_block}
\title{Create a materialized TDT block}
\usage{
new_tdt_block(info = list(), streams = list(), epocs = NULL)
}
\arguments{
\item{info}{Named list of block metadata.}

\item{streams}{Named list of stream objects.}

\item{epocs}{Epoc/event table. If \code{NULL}, an empty tibble is used.}
}
\value{
A \code{tdt_block} object.
}
\description{
Create a materialized TDT block
}
