% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_from_epocs}
\alias{ranges_from_epocs}
\title{Build time ranges from epoc/event onsets}
\usage{
ranges_from_epocs(epocs, pre, post, onset_col = "onset", drop_negative = TRUE)
}
\arguments{
\item{epocs}{A data frame with an onset column.}

\item{pre}{Start offset in seconds relative to onset. Use a negative value for
time before onset.}

\item{post}{Stop offset in seconds relative to onset.}

\item{onset_col}{Name of the onset column.}

\item{drop_negative}{If \code{TRUE}, drop ranges with negative start times.}
}
\value{
A numeric \verb{2 x N} matrix.
}
\description{
Build time ranges from epoc/event onsets
}
