% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{read_sev_py}
\alias{read_sev_py}
\title{Read TDT SEV data through Python \code{tdt}}
\usage{
read_sev_py(
  sev_dir,
  channel = 0,
  event_name = "",
  t1 = 0,
  t2 = 0,
  fs = 0,
  ranges = NULL,
  verbose = 0,
  just_names = FALSE,
  export = NULL,
  scale = 1,
  dtype = NULL,
  outdir = NULL,
  prefix = NULL,
  ...
)
}
\arguments{
\item{sev_dir}{Path to a SEV directory or file.}

\item{channel, event_name, t1, t2, fs, ranges, verbose, just_names, export, scale, dtype, outdir, prefix}{Arguments passed to Python \code{tdt.read_sev()}.}

\item{...}{Additional keyword arguments passed to Python \code{tdt.read_sev()}.}
}
\value{
A \code{tdt_sev_py} wrapper.
}
\description{
Read TDT SEV data through Python \code{tdt}
}
