% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{stream_names}
\alias{stream_names}
\title{Return stream names}
\usage{
stream_names(x)
}
\arguments{
\item{x}{A \code{tdt_block}, \code{tdt_block_py}, \code{tdt_sev_py}, stream container, or
compatible object.}
}
\value{
A character vector of stream names.
}
\description{
Return stream names
}
