% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{tdt_available}
\alias{tdt_available}
\title{Check whether Python \code{tdt} is available}
\usage{
tdt_available(initialize = TRUE)
}
\arguments{
\item{initialize}{If \code{TRUE}, allow reticulate to initialize Python while
checking for \code{tdt}.}
}
\value{
\code{TRUE} if reticulate can find Python \code{tdt}, otherwise \code{FALSE}.
}
\description{
Check whether Python \code{tdt} is available
}
