% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{tdt_calc_timepoints}
\alias{tdt_calc_timepoints}
\title{Calculate stream timepoints}
\usage{
tdt_calc_timepoints(n_points, fs, num_of_points = 128)
}
\arguments{
\item{n_points}{Number of samples in the stream.}

\item{fs}{Sampling frequency.}

\item{num_of_points}{Number of samples per exported row.}
}
\value{
A numeric vector of row-start timepoints in seconds.
}
\description{
Calculate stream timepoints
}
