% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{tdt_config}
\alias{tdt_config}
\title{Show Python configuration for \code{tdtr}}
\usage{
tdt_config(initialize = FALSE)
}
\arguments{
\item{initialize}{If \code{TRUE}, include \code{reticulate::py_config()} details. This
may initialize Python.}
}
\value{
A list with reticulate, Python, and Python \code{tdt} availability fields.
}
\description{
Show Python configuration for \code{tdtr}
}
