% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt-python.R
\name{tdt_get_stream}
\alias{tdt_get_stream}
\title{Get one stream from a TDT block}
\usage{
tdt_get_stream(block, stream)
}
\arguments{
\item{block}{A TDT block object returned by \code{\link[=tdt_read_block]{tdt_read_block()}} or an
equivalent R list.}

\item{stream}{Name of the stream to retrieve.}
}
\value{
A stream object.
}
\description{
Get one stream from a TDT block
}
