% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tank.R
\name{tdt_parse_tank_names}
\alias{tdt_parse_tank_names}
\title{Parse multiple TDT tank directory names}
\usage{
tdt_parse_tank_names(tank_dirs, tz = "UTC")
}
\arguments{
\item{tank_dirs}{Character vector of tank directory paths.}

\item{tz}{Time zone used for parsed datetimes.}
}
\value{
A tibble with one row per tank.
}
\description{
Parse multiple TDT tank directory names
}
