% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{tdt_stream_label}
\alias{tdt_stream_label}
\title{Return the display name for a stream object}
\usage{
tdt_stream_label(stream)
}
\arguments{
\item{stream}{A stream object or a character stream name.}
}
\value{
A single character string.
}
\description{
Return the display name for a stream object
}
