% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt-python.R
\name{tdt_stream_names}
\alias{tdt_stream_names}
\title{List stream names in a TDT block}
\usage{
tdt_stream_names(block)
}
\arguments{
\item{block}{A TDT block object returned by \code{\link[=tdt_read_block]{tdt_read_block()}} or an
equivalent R list.}
}
\value{
A character vector of stream names.
}
\description{
List stream names in a TDT block
}
