% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tdtR-package}
\alias{tdtR}
\alias{tdtR-package}
\title{tdtR: Read and Export Tucker-Davis Technologies Data from R}
\description{
\code{tdtR} provides a small R interface for reading TDT tank/block data through
Python's \code{tdt} package and exporting selected streams for downstream analysis.
}
\author{
\strong{Maintainer}: John Lee \email{johnleenimh@gmail.com}

Authors:
\itemize{
  \item John Lee \email{johnleenimh@gmail.com}
}

}
\keyword{internal}
