% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{validate_tdt_block}
\alias{validate_tdt_block}
\title{Validate a materialized TDT block}
\usage{
validate_tdt_block(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a materialized TDT block
}
