% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt-python.R
\name{tdt_available}
\alias{tdt_available}
\title{Check whether Python's \code{tdt} module is available}
\usage{
tdt_available()
}
\value{
\code{TRUE} if \code{reticulate} can find the Python \code{tdt} module, otherwise
\code{FALSE}.
}
\description{
Check whether Python's \code{tdt} module is available
}
