% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tdtr-package}
\alias{tdtr}
\alias{tdtr-package}
\title{tdtr: Read Tucker-Davis Technologies Data in R}
\description{
\code{tdtr} provides an R interface for reading TDT tank/block data through
Python's \code{tdt} package with reticulate, including Python-backed wrappers and
explicit helpers for collecting data into ordinary R objects for downstream
analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nimh-dsst/tdtr}
  \item \url{https://nimh-dsst.r-universe.dev/tdtr}
  \item Report bugs at \url{https://github.com/nimh-dsst/tdtr/issues}
}

}
\author{
\strong{Maintainer}: John Lee \email{johnleenimh@gmail.com}

Authors:
\itemize{
  \item John Lee \email{johnleenimh@gmail.com}
}

}
\keyword{internal}
